+++ cmd/web/routes.go
@@ -21,7 +21,8 @@ const APILogbookDelete = "/logbook/delete"
 const APILogbookNight = "/logbook/night"
 
 const APIExport = "/export"
-const APIExportFormat = APIExport + "/{format}"
+const APIExportFormat = "/export/{format}"
+const APIExportRestoreDefaults = "/export/restore"
 
 const APIImport = "/import"
 const APIImportCreateBackup = "/import/backup/create"
@@ -104,6 +105,7 @@ func (app *application) routes() *chi.Mux {
 		server.Get(APIExport, app.HandlerExport)
 		server.Get(APIExportFormat, app.HandlerExportLogbook)
 		server.Post(APIExportFormat, app.HandlerExportSettingsSave)
+		server.Post(APIExportRestoreDefaults, app.HandlerExportRestoreDefaults)
 
 		// import
 		server.Get(APIImport, app.HandlerImport)